u8 to u1

+++ UI/components/color_bar.py
@@ -17,7 +17,7 @@ class ColorBar(PixelDisplay):
         if width % 100 != 0:
             raise Exception("<width> for ColorBar must be multiples of 100")
         super(ColorBar, self).__init__(width, height, x, y)
-        self.color_buffer = numpy.array([self.colors['black'] for i in range(100)], dtype='u8')
+        self.color_buffer = numpy.array([self.colors['black'] for i in range(100)], dtype='u1')
         self.m = width
         self.n = height
         self.k = self.m // 100
@@ -28,7 +28,7 @@ class ColorBar(PixelDisplay):
         self.process()
 
     def process(self):
-        np_arr = numpy.zeros((self.m, self.n, 3), dtype='u8')
+        np_arr = numpy.zeros((self.m, self.n, 3), dtype='u1')
         for i in range(self.k):
             for j in range(self.n):
                 np_arr[i:self.m:self.k, j] = self.color_buffer
